+++ server/models/task.go
@@ -17,7 +17,7 @@ type Task struct {
 	Description string      `json:"description"`
 	Status      string      `json:"status"`
 	Comments    []Comment   `gorm:"foreignKey:TaskID" json:"comments"`
-	Position    int         `json:"position"`
+	Position    float64     `json:"position"`
 	CreatorID   uint        `json:"creator_id"`
 	Creator     User        `gorm:"foreignKey:CreatorID" json:"creator"`
 	AssigneeID  uint        `json:"assignee_id"`
